src/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pandas>=2.2.2
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-search-documents==11.6.0
azure-ai-inference==1.0.0b6
azure-ai-projects==1.0.0b5
azure-storage-blob==12.19.0
openai==1.54.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
jinja2==3.1.4
python-multipart==0.0.12
orjson==3.10.7
pydantic==2.9.2
Pillow>=10.4.0
aiofiles>=23.2.0
httpx>=0.27.0
psutil>=5.9.0
prometheus-client>=0.19.0
gunicorn>=21.2.0
aiosignal>=1.3.0
